refactor: rename to zwop
