+++ .gitignore
@@ -7,8 +7,8 @@ migration
 clients.json
 zuliprc.json
 node_modules
-src/zulip_write_only_proxy/_version.py
-src/zulip_write_only_proxy/frontend/static
+src/zwop/_version.py
+src/zwop/frontend/static
 zuliprc
 
 # Byte-compiled / optimized / DLL files
