+++ Dockerfile
@@ -12,14 +12,14 @@ COPY --link ./package.json ./pnpm-lock.yaml ./tailwind.config.js ./
 
 RUN --mount=type=cache,id=pnpm,target=/pnpm/store pnpm install --frozen-lockfile
 
-COPY ./src/zulip_write_only_proxy/frontend/templates/ \
-  ./src/zulip_write_only_proxy/frontend/templates/
+COPY ./src/zwop/frontend/templates/ \
+  ./src/zwop/frontend/templates/
 
 RUN pnpm build
 
 ADD --link https://unpkg.com/htmx.org@1.9.10/dist/htmx.js \
   https://unpkg.com/htmx.org@1.9.10/dist/htmx.min.js \
-  ./src/zulip_write_only_proxy/frontend/static/
+  ./src/zwop/frontend/static/
 
 
 ## Server
@@ -48,14 +48,14 @@ RUN --mount=type=cache,target=/opt/uv-cache/ \
     --mount=type=bind,source=.git,target=/app/.git \
   uv sync --locked
 
-COPY --link --from=frontend /app/src/zulip_write_only_proxy/frontend \
-  /app/src/zulip_write_only_proxy/frontend
+COPY --link --from=frontend /app/src/zwop/frontend \
+  /app/src/zwop/frontend
 
 EXPOSE 8000
 
 ENV ZWOP_ADDRESS=http://0.0.0.0:8000
 
-CMD ["uv", "run", "-m", "zulip_write_only_proxy.main"]
+CMD ["uv", "run", "-m", "zwop.main"]
 
 HEALTHCHECK --start-interval=1s --start-period=30s --interval=60s \
   CMD wget http://0.0.0.0:8000/api/health || exit 1
