+++ package.json
@@ -4,9 +4,9 @@
     "tailwindcss": "^3.4.1"
   },
   "scripts": {
-    "htmx": "cd  src/zulip_write_only_proxy/frontend/static/; wget -N https://unpkg.com/browse/htmx.org@1.9.10/dist/htmx.js; wget -N https://unpkg.com/browse/htmx.org@1.9.10/dist/htmx.min.js",
-    "build": "tailwindcss -o src/zulip_write_only_proxy/frontend/static/main.css",
-    "watch": "tailwindcss -o src/zulip_write_only_proxy/frontend/static/main.css --watch"
+    "htmx": "cd  src/zwop/frontend/static/; wget -N https://unpkg.com/browse/htmx.org@1.9.10/dist/htmx.js; wget -N https://unpkg.com/browse/htmx.org@1.9.10/dist/htmx.min.js",
+    "build": "tailwindcss -o src/zwop/frontend/static/main.css",
+    "watch": "tailwindcss -o src/zwop/frontend/static/main.css --watch"
   },
   "devDependencies": {
     "prettier": "^3.2.5",
