+++ src/zwop/frontend/templates/fragments/list-table-rows.html
