+++ src/zwop/frontend/templates/fragments/navbar.html
