+++ src/zwop/frontend/templates/list.html
