+++ tailwind.config.js
@@ -1,8 +1,8 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
   content: [
-    "./src/zulip_write_only_proxy/frontend/templates/**.html",
-    "./src/zulip_write_only_proxy/frontend/templates/*/**.html"
+    "./src/zwop/frontend/templates/**.html",
+    "./src/zwop/frontend/templates/*/**.html"
   ],
   theme: {
     extend: {},
